Capitalise TrustedInstaller & PowerShell for RunAsTI
Windows Terminal has never worked for RunAsTI in my experience. On Windows 10, it would not connect to the Touch Keyboard and Handwriting Panel Service properly, which made it so you could not type. On Windows 11, it seems to just not launch at all. Using pure Console Host and PowerShell seems like the more reliable option.

Fixes typos: "Run as trustedinstaller" -> "Run as TrustedInstaller"

Fixes changelog being not on a new line in the README.md